requirements.txt
requests
aiohttp
websockets
pandas
python-dotenv
schedule
